chore: move notifier into `/internal`
